refactor: 서버컴포넌트/클라컴포넌트 이름 정리
