+++ src/app/(default)/album/_client-album-page.tsx
